requirements.txt
# Firecrawl 测试所需依赖
firecrawl-py>=0.1.0
pandas>=1.3.0
openpyxl>=3.0.9 